# docs/swagger.yaml
basePath: /api/
definitions:
...
  models.User:
    properties:
      icon:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - icon
    - password
    - username
    type: object
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  users.IdResponse:
    properties:
      id:
        type: string
    type: object
  users.ListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  users.MessageResponse:
    properties:
      message:
        type: string
    type: object
  users.StatusesListResponse:
    properties:
      blacklist:
        items:
          $ref: '#/definitions/models.User'
        type: array
      friends:
        items:
          $ref: '#/definitions/models.User'
        type: array
      invites:
        items:
          $ref: '#/definitions/models.User'
        type: array
      onBlacklist:
        items:
          $ref: '#/definitions/models.User'
        type: array
      requires:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  users.UserResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
host: localhost:8000
info:
  contact:
    name: Tairyoku
    url: https://github.com/Tairyoku
  description: Серверна частина онлайн чату.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Server API
  version: 1.0.0
paths:
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Повертає дані користувача.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: return user`s data
          schema:
            $ref: '#/definitions/users.UserResponse'
        "500":
          description: get user error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user`s data by ID
      tags:
      - users
  /users/{id}/accept:
    put:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Підтверджує запит на дружбу поточного користувача з отриманим.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: invitation accepted
          schema:
            $ref: '#/definitions/users.MessageResponse'
        "500":
          description: update status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Accept friendship invitation
      tags:
      - users
  /users/{id}/addToBL:
    post:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Блокує користувача.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: user is blocked
          schema:
            $ref: '#/definitions/users.IdResponse'
        "500":
          description: add status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Block user
      tags:
      - users
  /users/{id}/all:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Повертає списки відносин між користувачем та іншими користувачами.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: return user`s status lists
          schema:
            $ref: '#/definitions/users.StatusesListResponse'
        "500":
          description: friend requires list error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user`s relationship lists by ID
      tags:
      - users
  /users/{id}/cancel:
    delete:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Видаляє запит на дружбу поточного користувача з отриманим.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: invite deleted
          schema:
            $ref: '#/definitions/users.MessageResponse'
        "500":
          description: delete status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Cancel friendship invitation
      tags:
      - users
  /users/{id}/deleteFriend:
    delete:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Видаляє користувача зі списку друзів.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: friend deleted
          schema:
            $ref: '#/definitions/users.MessageResponse'
        "500":
          description: delete status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete friend
      tags:
      - users
  /users/{id}/deleteFromBlacklist:
    delete:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Видаляє користувача зі списку заблокованих.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: user deleted from black list
          schema:
            $ref: '#/definitions/users.MessageResponse'
        "500":
          description: delete status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete from blacklist
      tags:
      - users
  /users/{id}/invite:
    post:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Створює запит на дружбу поточного користувача з отриманим.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: require is sent
          schema:
            $ref: '#/definitions/users.IdResponse'
        "500":
          description: add status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create friendship invitation
      tags:
      - users
  /users/{id}/refuse:
    put:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Відхиляє запит на дружбу поточного користувача з отриманим.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: invitation refused
          schema:
            $ref: '#/definitions/users.MessageResponse'
        "500":
          description: delete status error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refuse friendship invitation
      tags:
      - users
  /users/search/{username}:
    get:
      consumes:
      - application/json
      description: |-
        Отримує частину імені чату.
        Повертає список користувачів, ім'я яких повністю або частково збігається.
      parameters:
      - description: Slice of users
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: lisl of found chats
          schema:
            $ref: '#/definitions/users.ListResponse'
        "500":
          description: search users error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get found chats
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"